+++ RELEASES.md
@@ -657,3 +657,8 @@
 -----------------
   - listUsers and listGroups now allow for extended search options (sort, limit, skip, count)
   - updates to data provider to allow for aggregate, collation and count in mongodb
+
+10.4.1 2019-10-08
+-----------------
+  - listUsers collation fix
+  - mongodb listUsers and listGroups tests